pipeline/requirements.txt
pandas==2.2.3
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-genai>=1.0.0
openai==1.58.1
pydantic>=2.7.0
requests==2.32.3
ruff==0.8.6
pytest==8.3.4

